move LoopingCallManager to own module
